feat: update command type (VF-000) (#265)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements VF-000**

### Brief description. What is this change?
To accompany @astopo 's migration. It turns out many legacy google and alexa projects (as mentioned in the comments) either don't have a `command[].type` or they are just `type: "intent"` or `type: "command"`.

The new format we use is either `"jump"` or `"push"` which will bring it in line with `general-runtime`. During the migration we will convert ALL command types to this new format. What is nice is the runtime doesn't care about the actual `type` property at all, it just checks for the existence of other properties, so this is non-breaking downstream.

google-runtime:
https://github.com/voiceflow/google-runtime/blob/f4cdb47be45df8c61a076e46a17e59a6d7e9d6fb/lib/services/runtime/handlers/command.ts#L12-L18

alexa-runtime:
https://github.com/voiceflow/alexa-runtime/blob/74a292aef66964b0d7a3c6161b7dd6dcfa2d3c45/lib/services/runtime/handlers/command.ts#L14-L20