fix: reference types in published directory (VF-000) (#329)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements VF-000**

### Brief description. What is this change?

The `package.json` includes a `files` property that is used to say what is included in the published tarball. These all reference the `build` directory: the output of the typescript compiler.
There is also a `types` property that is used to tell typescript where the types are located for that package. These were referencing the `src` directory, which is not included in the published package. Therefore typescript could not find types, since they reference to them didn't exist.

Depends on: https://github.com/voiceflow/tsconfig/pull/30

Co-authored-by: Tyler Stewart <tyler.stewart@voiceflow.com>
Co-authored-by: Ben Teichman <ben@voiceflow.com>