feat: add base publishing property (CT-000) (#371)
As part of my recent dialogflowCX work, I've had to add a `selectedIntents` field under `version.platformData.publishing`
This data needs to be persisted so we can configure it from different areas of the app.

The publishing job will read this value and use it to decide which intents to upload. We can see it here as part of the dialogflowCX publishing process:
![Screen Shot 2022-09-27 at 5 09 20 PM](https://user-images.githubusercontent.com/5643574/192635976-0751f3bc-b9de-4433-88ac-85e456ad7c5e.png)
https://github.com/voiceflow/google-service/blob/1bc9a877afc784d7d5140cfd024005b10b0c8668/lib/jobs/dialogflow/cx/publish/stages/agent.ts#L168-L173

This might be a requirement for additional NLPs we upload to in the future. Right now publishing is just `Record<string, any>` so this will help it be more definitively typed.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>