feat: upload custom tls certificate for api calls (PL-120) (#373)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements PL-120**

### Brief description. What is this change?
Adds an optional `tls` field to `StepData` and `NodeData`. The `tls` field contains a `cert` and `key` field which are both optional strings.
<!-- Build up some context for your teammates on the changes made here and potential tradeoffs made and/or highlight any topics for discussion -->
This change is part of PL-120. It allows the key name of a certificate and key to be stored on a node and then used to retrieve the actual file at runtime.