feat: version kb (nlu-906) (#460)
Add ability to version the KB. The goal is to have kb settings related to the version and not the project. I used the full knowledgeBase model which includes support for docs (even though we won't be using them yet) to avoid having to backtrack changes in the future when we want to version KB docs along with settings.