feat: adding serialized check (CV3-289) (#461)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements CV3-289**

### Brief description. What is this change?

<!-- Build up some context for your teammates on the changes made here and potential tradeoffs made and/or highlight any topics for discussion -->

Adding a new static method to `HTTPException` to check if a plain object is an instance of a serialized `HTTPException`.

The reason this method should be added as part of the `HTTPException` API is because the `HTTPException` is often serialized and transmitted over the wire to a different service. The client service which receives the serialized `HTTPException` needs a method to detect that the error payload is of shape `SerializedHTTPException` and not something else like a basic JavaScript `Error`. This is a very common use-case. 

Without this method, every client has to implement the duck-typing check for `SerializedHTTPException` to proper handle this exception. This creates duplication.